Implement support for other Jenkins hosts
Adds an initial attempt at allowing the extension to support non-dotnet Jenkins hosts.

To avoid hardcoding the URLs of other hosts we need to add an "optional_permissions" key to manifest.json.
By claiming we need http://*/* we can request permissions for specific URLs on demand.

Unfortunately, the chrome.permissions.request API requires being invoked from an explicit user action
like a button click.

To add to that we can't even access it from a button in the content.js script, it only seems to work
in options.js. Googling revealed a few workarounds involving an iframe that sends messages to
options.js, but that seems too involved for now.

Instead, I opted for the simplest solution of providing a text box in options.html where you can
simply paste a new URL, press the button and Chrome will ask for confirmation.

Fixes #57